web-dashboard-service/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
plotly>=5.15.0
numpy>=1.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
wordcloud>=1.9.0
matplotlib>=3.7.0

sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL 연결을 위한 드라이버
python-dotenv>=1.0.0  # 환경변수 로딩